+++ nullaway/src/main/java/com/uber/nullaway/handlers/OptionalEmptinessHandler.java
@@ -52,9 +52,11 @@ public class OptionalEmptinessHandler extends BaseNoOpHandler {
 
   @Nullable private ImmutableSet<Type> optionalTypes;
   private final Config config;
+  private final MethodNameUtil methodNameUtil;
 
-  OptionalEmptinessHandler(Config config) {
+  OptionalEmptinessHandler(Config config, MethodNameUtil methodNameUtil) {
     this.config = config;
+    this.methodNameUtil = methodNameUtil;
   }
 
   @Override
@@ -94,15 +96,11 @@ public class OptionalEmptinessHandler extends BaseNoOpHandler {
     if (optionalIsGetCall(symbol, types)) {
       return NullnessHint.HINT_NULLABLE;
     } else if (optionalIsPresentCall(symbol, types)) {
-      Element getter = null;
-      Node base = node.getTarget().getReceiver();
-      for (Symbol elem : symbol.owner.getEnclosedElements()) {
-        if (elem.getKind().equals(ElementKind.METHOD)
-            && elem.getSimpleName().toString().equals("get")) {
-          getter = elem;
-        }
-      }
-      updateNonNullAPsForElement(thenUpdates, getter, base);
+      updateNonNullAPsForElement(
+          thenUpdates, getOptionalGetter(symbol), node.getTarget().getReceiver());
+    } else if (config.handleTestAssertionLibraries() && methodNameUtil.isMethodIsTrue(symbol)) {
+      // we check for instance of AssertThat(optionalFoo.isPresent()).isTrue()
+      updateIfAssertIsPresentTrueOnOptional(node, types, bothUpdates);
     }
     return NullnessHint.UNKNOWN;
   }
@@ -139,6 +137,33 @@ public class OptionalEmptinessHandler extends BaseNoOpHandler {
     return false;
   }
 
+  private void updateIfAssertIsPresentTrueOnOptional(
+      MethodInvocationNode node, Types types, AccessPathNullnessPropagation.Updates bothUpdates) {
+    Node receiver = node.getTarget().getReceiver();
+    if (receiver instanceof MethodInvocationNode) {
+      MethodInvocationNode receiverMethod = (MethodInvocationNode) receiver;
+      Symbol.MethodSymbol receiverSymbol = ASTHelpers.getSymbol(receiverMethod.getTree());
+      if (methodNameUtil.isMethodAssertThat(receiverSymbol)) {
+        // assertThat will always have at least one argument, So safe to extract from the arguments
+        Node arg = receiverMethod.getArgument(0);
+        if (arg instanceof MethodInvocationNode) {
+          // Since assertThat(a.isPresent()) changes to
+          // Truth.assertThat(Boolean.valueOf(a.isPresent()))
+          // need to be unwrapped from Boolean.valueOf
+          Node unwrappedArg = ((MethodInvocationNode) arg).getArgument(0);
+          if (unwrappedArg instanceof MethodInvocationNode) {
+            MethodInvocationNode argMethod = (MethodInvocationNode) unwrappedArg;
+            Symbol.MethodSymbol argSymbol = ASTHelpers.getSymbol(argMethod.getTree());
+            if (optionalIsPresentCall(argSymbol, types)) {
+              updateNonNullAPsForElement(
+                  bothUpdates, getOptionalGetter(argSymbol), argMethod.getTarget().getReceiver());
+            }
+          }
+        }
+      }
+    }
+  }
+
   private void updateNonNullAPsForElement(
       AccessPathNullnessPropagation.Updates updates, @Nullable Element elem, Node base) {
     if (elem != null) {
@@ -147,6 +172,16 @@ public class OptionalEmptinessHandler extends BaseNoOpHandler {
         updates.set(ap, Nullness.NONNULL);
       }
     }
+  }
+
+  private Element getOptionalGetter(Symbol.MethodSymbol symbol) {
+    for (Symbol elem : symbol.owner.getEnclosedElements()) {
+      if (elem.getKind().equals(ElementKind.METHOD)
+          && elem.getSimpleName().toString().equals("get")) {
+        return elem;
+      }
+    }
+    return null;
   }
 
   private boolean optionalIsPresentCall(Symbol.MethodSymbol symbol, Types types) {