+++ nullaway/src/test/resources/com/uber/nullaway/testdata/androidstubs/androidx/Fragment.java
@@ -40,4 +40,6 @@ public class Fragment {
       LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
     return null;
   }
+
+  public void onActivityCreated(Bundle savedInstanceState) {}
 }